+++ packages/nimble-components/src/table/components/cell/tests/table-cell.pageobject.ts
@@ -1,4 +1,5 @@
 import type { TableCell } from '..';
+import { TableCellView } from '../../../../table-column/base/cell-view';
 import type { TableCellRecord } from '../../../../table-column/base/types';
 
 /**
@@ -8,11 +9,17 @@ import type { TableCellRecord } from '../../../../table-column/base/types';
 export class TableCellPageObject<T extends TableCellRecord = TableCellRecord> {
     public constructor(private readonly tableCellElement: TableCell<T>) {}
 
-    public getRenderedCellContent(): Element | undefined {
-        const cellContent = this.tableCellElement.shadowRoot!.querySelector(
-            '.cell-content-container'
-        );
+    public getRenderedCellView(): TableCellView {
+        const cellView = this.tableCellElement.shadowRoot!.firstElementChild;
+        if (!(cellView instanceof TableCellView)) {
+            throw new Error(
+                'Cell view not found in cell - ensure cellViewTag is set for column'
+            );
+        }
+        return cellView as TableCellView;
+    }
 
-        return cellContent?.children[0];
+    public getRenderedCellContent(): string {
+        return this.getRenderedCellView().shadowRoot!.textContent?.trim() ?? '';
     }
 }