+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,42 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Mon, 03 Apr 2023 21:31:32 GMT and should not be manually modified.
+This log was last generated on Wed, 05 Apr 2023 18:54:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 16.5.1
+
+Wed, 05 Apr 2023 18:54:28 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v18.10.1
+
+## 16.5.0
+
+Wed, 05 Apr 2023 17:16:16 GMT
+
+### Minor changes
+
+- Angular support for anchor tree item ([ni/nimble@1827deb](https://github.com/ni/nimble/commit/1827debb2b303a140bee1b23a01f8d94dbc35247))
+- Bump @ni/nimble-components to v18.10.0
+
+## 16.4.1
+
+Tue, 04 Apr 2023 23:23:52 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v18.9.1
+
+## 16.4.0
+
+Tue, 04 Apr 2023 19:42:20 GMT
+
+### Minor changes
+
+- Error state for text area component ([ni/nimble@32bd2ad](https://github.com/ni/nimble/commit/32bd2adecfea401853993df53e8307e060ceb45a))
+- Bump @ni/nimble-components to v18.9.0
 
 ## 16.3.0